Migrate WritePoint to unsafe native pointer and split LasPoint.cs
- Extract LasZipPointNative to its own file with new WriteFromLasPoint method
- Extract LasZipPointStruct to its own file, marked [Obsolete]
- Rewrite WritePoint to use LasZipPointNative* instead of Marshal.StructureToPtr
- Remove legacy ConvertPoint(LasZipPointStruct) overload from LasPoint
- Fix bug where old WritePoint mutated caller's point coordinates

Closes #11
